Patch #8177: Timeline Performance Patch Patch by: Adriano Macchietto (macchiea)
This patch optimises the way keyframes are drawn in the Timeline, so that it is more responsive when working with heaps of keyframes (i.e. motion-capture data).

Detailed Description of Patch (from author):
* No longer uses a list to store the keys before drawing. Uses
less memory.
* Culls the drawing of keyframes outside of the visible window.
Good for dealing with long mocap tracks.
* Performs a check to avoid redrawing a line over a line which
has already been drawn to the same pixels. This speeds up the
scenario when you have many keyframes and are zoomed out.
* Batches the draws into one glBegin/glEnd block.

This is all done on a per IpoCurve basis.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13510 954f8c5b-7b00-dc11-b283-0030488c597c