== Sculpt ==
Added a new brush option, "Anchored". When enabled, the brush doesn't move with the mouse, but rather stays in it's initial location and grows larger or smaller to follow the mouse. Good for brushing alphas on to the mesh. (Note that this option isn't available for the grab brush, and ignores the smooth stroke option.)



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@13515 954f8c5b-7b00-dc11-b283-0030488c597c