Potential fix for [#8113] Blender.Registry segfault when no permission to write
This converts exceptions in Warning in selected points of the registry module.

While I couldn't reproduce the segfault ( I received the Py error as expected), this isn't a bad idea anyway, since the BPy C counterpart was discarding the error anyway, making it a warning is a bit more inline with its meaning.

The exception is not preventable nor fixable from a script's point of view, so better to silence it and print a warning to the console instead.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14353 954f8c5b-7b00-dc11-b283-0030488c597c