Changed BLI_convertstringcode to replace any number of hashes with the frame number.
somefile_##.png -> somefile_01.png
somefile_########-image.png -> somefile_00000001-image.png

Before, A hash at the end of the string would be replaced by a number with 4 characters. This is still default if no #'s are in the string, so nothing has changed.

To use this function from the python api use scene.render.getFrameFilename()


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14385 954f8c5b-7b00-dc11-b283-0030488c597c