== Action Editor - Group Colours for Channels ==
It is now possible for Action Channels to be displayed using the colour of its group. The colour for an Action Group is only set when it is auto-created by keyframing (controls to tweak this may/may-not be added, time permitting). 

To enable the display of these colours, the View -> 'Use Group Colors' must be enabled. However, this is disabled by default.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14438 954f8c5b-7b00-dc11-b283-0030488c597c