Reverting vrml97_export.py to it's state at revision 14751 because:
1) Bug #6692 is actually a feature request. While I do think that
Michalis' patch is a worthy change, we don't have time to test
this at this stage in the release cycle (and svn is supposedly
closed to new features anyways). I'm sorry Michalis, this will
have to wait until after release.

2) I have looked at bug #8814 a while ago, but I have not yet found
a good solution (and perhaps the solution is to *not* export faces 
without materials assigned). I don't have time to assess whether
revision 14774 presents a worthwhile change or not, and I have no
time to test before release. Also I don't like that the 'fix' for
#8814 is mixed with a fix for a previous faulty commit.

3) Again, another commit to a script I maintain without consulting
me first. Not cool!

Chris



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@14779 954f8c5b-7b00-dc11-b283-0030488c597c