Fix segfault in the Sequence.
If you enable the Proxy and Custom Dir options and later disable
the Proxy (but not the Custom Dir options), Blender crash every
time that you try select the strip, because the proxy don't exist
anymore.

The solution: only draw the Custom Dir button if the Proxy
option is enable.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@15579 954f8c5b-7b00-dc11-b283-0030488c597c