Bugfix #14428: Constraints lost when object made single user
When there was a setup where an object was linked to more than one scene at once, and in one of those scenes some of the objects in that scene were related to it, the objects related to it were not correctly relinked to the new copy. 

This was due to the 2nd check for selected objects 
if( (base->flag & flag)==flag)
meaning that only selected objects would get corrected.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@16674 954f8c5b-7b00-dc11-b283-0030488c597c