Added getitem/setitem access for KX_GameObject ob.someProp = 10 can now be... ob["someProp"] = 10
For simple get/set test with an objects 10 properties, this is ~30% faster.

Though I like the attribute access, its slower because it needs to lookup BGE attributes and methods (for parent classes as well as KX_GameObject class).

This could also be an advantage if there are collisions between new attributes added for 2.49 and existing properties a game uses.

Made some other small optimizations,
- Getting and setting property can use const char* as well as STR_String (avoids making new STR_Strings just to do the lookup).
- CValue::SetPropertiesModified() and CValue::SetPropertiesModified(), were looping through all items in the std::map, advancing from the beginning each time.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@19497 954f8c5b-7b00-dc11-b283-0030488c597c